# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de e-commerce",
        "title": "API de E-commerce",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.suporte.com",
            "email": "suporte@api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pedidos": {
            "get": {
                "description": "Retorna a lista completa de pedidos cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Lista todos os pedidos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Pedido"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo pedido no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Adiciona um novo pedido",
                "parameters": [
                    {
                        "description": "Dados do Pedido",
                        "name": "pedido",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Pedido"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Cliente ou produto não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Estoque insuficiente",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/pedidos/search": {
            "get": {
                "description": "Retorna os pedidos cujos clientes têm nomes que correspondem ao parâmetro de busca",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Busca pedidos por nome do cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome ou parte do nome do cliente para busca",
                        "name": "nome",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Pedido"
                            }
                        }
                    },
                    "400": {
                        "description": "Nome não pode ser vazio",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pedidos/{id}": {
            "get": {
                "description": "Retorna os detalhes de um pedido específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Busca um pedido por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pedido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Pedido"
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um pedido do sistema (apenas pedidos cancelados)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Remove um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pedido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Pedido não pode ser deletado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pedidos/{id}/cancelar": {
            "post": {
                "description": "Cancela um pedido e devolve os produtos ao estoque",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Cancela um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pedido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Pedido não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pedidos/{id}/status": {
            "put": {
                "description": "Altera o status de um pedido existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Atualiza status do pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pedido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novo status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Status inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ItemPedido": {
            "type": "object",
            "properties": {
                "preco_unit": {
                    "type": "number"
                },
                "produto_id": {
                    "type": "string"
                },
                "quantidade": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "number"
                }
            }
        },
        "model.Pedido": {
            "type": "object",
            "properties": {
                "cliente_id": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "itens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ItemPedido"
                    }
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
    }
}